Fixed an issue with the clone provider
- Fixed an issue with the clone provider keep_running
- Added a error management for when cloning a running vm
- Added the keep_after_error to allow vms to stay if there is an error

+++ internal/remoteimage/resource.go
@@ -751,7 +751,9 @@ func (r *RemoteVmResource) Update(ctx context.Context, req resource.UpdateReques
 					resp.Diagnostics.Append(diag...)
 					return
 				}
-				apiclient.DeleteVm(aptCtx, hostConfig, data.ID.ValueString())
+        if !data.KeepAfterError.ValueBool() {
+				  apiclient.DeleteVm(aptCtx, hostConfig, data.ID.ValueString())
+        }
 			}
 			return
 		}
@@ -1103,7 +1105,9 @@ func updateReverseProxyHostsTarget(ctx context.Context, data *models.RemoteVmRes
 			if _, diag := common.EnsureMachineStopped(ctx, hostConfig, refreshedVm); diag.HasError() {
 				return nil, diag
 			}
-			apiclient.DeleteVm(ctx, hostConfig, data.ID.ValueString())
+      if !data.KeepAfterError.ValueBool() {
+			  apiclient.DeleteVm(ctx, hostConfig, data.ID.ValueString())
+      }
 		}
 		return nil, resultDiagnostic
 	}
